backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[dotenv]
gitpython
python-dotenv
pytest
